Add files via upload
I've made changes to pymgrid-0.1.0-beta to introduce variability into each reinforcement learning training episode. The author of these changes is __.
The modified methods are __init__, reset, update_variables, and run.
The main purpose of these changes is to allow the reset method to randomize the battery's initial state of charge (SoC) and randomly select a starting point within the time series (charge and PV).