Add files via upload
I've made chenges to pymgrid-0.1.0-beta  to introduce variability into each episode in
Reinforcement Learning training. The author of these changes is Ana Isabel Ascencio Pedraza
The methods with changes are __init__, reset, update_variables and run. 
The main objective of this changes was  to posibilitate that the reset method randomize 
the battery's initial State of Charge and randomly select a starting point whithin the long
time series (load and pv)